Add shared folder and move Modal to shared. Block scrolling of main page when scrolling modal.
