+++ client/source/components/shared/Modal.jsx
@@ -52,6 +52,12 @@ const CloseButton = styled.button`
 `;
 
 export default function Modal({ children, isOpen, onClose }) {
+  if (isOpen) {
+    document.body.classList.add('stop-scrolling');
+  } else {
+    document.body.classList.remove('stop-scrolling');
+  }
+
   return (
     isOpen && createPortal(
       <>
